SQL.Table: default generic implementation of createTable
