+++ kore/src/Kore/Log/WarnBottomHook.hs
@@ -54,7 +54,6 @@ instance Entry WarnBottomHook where
     entrySeverity _ = Warning
 
 instance SQL.Table WarnBottomHook where
-    createTable = SQL.createTableGeneric
     insertRow = SQL.insertRowGeneric
     selectRow = SQL.selectRowGeneric
